start boa scripts for deployments
