+++ deployments.yaml
@@ -1,0 +1,12 @@
+ethereum:mainnet:
+  factory: "0x0c0e5f2fF0ff18a3be9b835635039256dC4B4963"
+  math: "0xcBFf3004a20dBfE2731543AA38599A526e0fD6eE"
+  views: "0x064253915b8449fdEFac2c4A74aA9fdF56691a31"
+  amm: "0x66442B0C5260B92cAa9c234ECf2408CBf6b19a6f"
+  gauge: "0x5fC124a161d888893529f67580ef94C2784e9233"
+  factory_handler: "0x30a4249C42be05215b6063691949710592859697"
+arbitrum:mainnet:
+  factory: "0xbC0797015fcFc47d9C1856639CaE50D0e69FbEE8"
+  math: "0x604388Bb1159AFd21eB5191cE22b4DeCdEE2Ae22"
+  views: "0x06452f9c013fc37169B57Eab8F50A7A48c9198A3"
+  amm: "0xd7E72f3615aa65b92A4DBdC211E296a35512988B"
