+++ scripts/deployment_utils.py
@@ -1,178 +1,119 @@
-import math
+# flake8: noqa E501
+
+import sys
 from dataclasses import dataclass
 
-import click
-from ape import Contract, networks, project
-from ape.api.address import Address
-from ape.logging import logger
-from eth_abi import encode
-from pycoingecko import CoinGeckoAPI
+from eth_typing import Address
+from eth_utils import keccak
+from rich.console import Console as RichConsole
 
-DOLLAR_VALUE_OF_TOKENS_TO_DEPOSIT = 5
+logger = RichConsole(file=sys.stdout)
 
-
-def _get_tx_params():
-
-    if "mainnet-fork" == networks.active_provider.network.name:
-        return {}
-
-    if "sepolia" == networks.active_provider.network.name:
-        return {}
-
-    active_provider = networks.active_provider
-    max_fee = int(active_provider.base_fee * 1.2)
-    max_priority_fee = int(0.5e9)
-
-    return {"max_fee": max_fee, "max_priority_fee": max_priority_fee}
-
-
-def deploy_blueprint(contract, account):
-
-    initcode = contract.contract_type.deployment_bytecode.bytecode
-    if isinstance(initcode, str):
-        initcode = bytes.fromhex(initcode.removeprefix("0x"))
-    initcode = b"\xfe\x71\x00" + initcode  # eip-5202 preamble version 0
-    initcode = (
-        b"\x61"
-        + len(initcode).to_bytes(2, "big")
-        + b"\x3d\x81\x60\x0a\x3d\x39\xf3"
-        + initcode
-    )
-
-    tx = project.provider.network.ecosystem.create_transaction(
-        chain_id=project.provider.chain_id,
-        data=initcode,
-        gas_price=project.provider.gas_price,
-        nonce=account.nonce,
-        **_get_tx_params(),
-    )
-    receipt = account.call(tx)
-    click.echo(f"blueprint deployed at: {receipt.contract_address}")
-    return receipt.contract_address
-
-
-def get_deposit_amounts(amount_per_token_usd, initial_prices, coins):
-    initial_prices = [10**18] + initial_prices
-    precisions = [
-        10 ** project.ERC20Mock.at(coin).decimals() for coin in coins
-    ]
-    deposit_amounts = [
-        amount_per_token_usd * precision * 10**18 // price
-        for price, precision in zip(initial_prices, precisions)
-    ]
-    return deposit_amounts
-
-
-# -------------- CURVE DATA --------------
+GAUGE_CONTROLLER = "0x2F50D538606Fa9EDD2B11E2446BEb18C9D5846bB"
+ADDRESS_PROVIDER = "0x0000000022d53366457f9d5e68ec105046fc4383"
+FIDDYRESEARCH = "0xE6DA683076b7eD6ce7eC972f21Eb8F91e9137a17"
+FIDDYDEPLOYER = "0x2d12D0907A388811e3AA855A550F959501d303EE"
+BABE = "0xbabe61887f1de2713c6f97e567623453d3C79f67"
 
 
 @dataclass
 class CurveNetworkSettings:
-    dao_ownership_contract: Address
-    fee_receiver_address: Address
-    usdc_address: Address
-    wbtc_address: Address
-    weth_address: Address
-
-
-GAUGE_CONTROLLER = "0x2F50D538606Fa9EDD2B11E2446BEb18C9D5846bB"
-ADDRESS_PROVIDER = "0x0000000022d53366457f9d5e68ec105046fc4383"
+    dao_ownership_contract: Address = ""
+    thin_proxy_contract: Address = ""
+    fee_receiver_address: Address = ""
+    metaregistry_address: Address = ""
+    base_pool_registry_address: Address = ""
+    address_provider: Address = "0x0000000022d53366457f9d5e68ec105046fc4383"
 
 
 curve_dao_network_settings = {
-    "ethereum:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="0x40907540d8a6C65c637785e8f8B742ae6b0b9968",
-        fee_receiver_address="0xeCb456EA5365865EbAb8a2661B0c503410e9B347",
-        usdc_address="0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-        wbtc_address="0x2260fac5e5542a773aa44fbcfedf7c193bc2c599",
-        weth_address="0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
-    ),
+    # Ethereum
     "ethereum:sepolia": CurveNetworkSettings(
         dao_ownership_contract="0xE6DA683076b7eD6ce7eC972f21Eb8F91e9137a17",
         fee_receiver_address="0xE6DA683076b7eD6ce7eC972f21Eb8F91e9137a17",
-        usdc_address="0x51fCe89b9f6D4c530698f181167043e1bB4abf89",
-        wbtc_address="0xFF82bB6DB46Ad45F017e2Dfb478102C7671B13b3",
-        weth_address="0xf531B8F309Be94191af87605CfBf600D71C2cFe0",
     ),
+    "ethereum:mainnet": CurveNetworkSettings(
+        dao_ownership_contract="0x40907540d8a6C65c637785e8f8B742ae6b0b9968",
+        thin_proxy_contract="0xEdf2C58E16Cc606Da1977e79E1e69e79C54fe242",
+        fee_receiver_address="0xeCb456EA5365865EbAb8a2661B0c503410e9B347",  # burner
+        metaregistry_address="0xF98B45FA17DE75FB1aD0e7aFD971b0ca00e379fC",
+        base_pool_registry_address="0xDE3eAD9B2145bBA2EB74007e58ED07308716B725",
+    ),
+    # Layer 2
     "arbitrum:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="0xb055ebbacc8eefc166c169e9ce2886d0406ab49b",
-        fee_receiver_address="0xd4f94d0aaa640bbb72b5eec2d85f6d114d81a88e",
-        usdc_address="0xaf88d065e77c8cC2239327C5EDb3A432268e5831",  # native usdc token.  # noqa: E501
-        wbtc_address="0x2f2a2543b76a4166549f7aab2e75bef0aefc5b0f",
-        weth_address="0x82aF49447D8a07e3bd95BD0d56f35241523fBab1",
+        dao_ownership_contract="0x452030a5D962d37D97A9D65487663cD5fd9C2B32",
+        thin_proxy_contract="0xb055ebbacc8eefc166c169e9ce2886d0406ab49b",
+        fee_receiver_address="0xd4F94D0aaa640BBb72b5EEc2D85F6D114D81a88E",  # burner
     ),
     "optimism:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
-        fee_receiver_address="0xbF7E49483881C76487b0989CD7d9A8239B20CA41",
-        usdc_address="0x7f5c764cbc14f9669b88837ca1490cca17c31607",
-        wbtc_address="0x68f180fcce6836688e9084f035309e29bf0a2095",
-        weth_address="0x4200000000000000000000000000000000000006",
+        dao_ownership_contract="0x28c4A1Fa47EEE9226F8dE7D6AF0a41C62Ca98267",
+        thin_proxy_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
+        fee_receiver_address="0xbF7E49483881C76487b0989CD7d9A8239B20CA41",  # burner
     ),
+    "base:mainnet": CurveNetworkSettings(
+        dao_ownership_contract="0x2c163fe0f079d138b9c04f780d735289344C8B80",
+        thin_proxy_contract="0xe8269B33E47761f552E1a3070119560d5fa8bBD6",
+        fee_receiver_address="0xe8269B33E47761f552E1a3070119560d5fa8bBD6",
+    ),
+    "linea:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
+        fee_receiver_address="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
+    ),
+    "scroll:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
+        fee_receiver_address="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
+    ),
+    "pzkevm:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0x8b3EFBEfa6eD222077455d6f0DCdA3bF4f3F57A6",
+        fee_receiver_address="0x8b3EFBEfa6eD222077455d6f0DCdA3bF4f3F57A6",
+    ),
+    # Layer 1
     "polygon:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
-        fee_receiver_address="0x774D1Dba98cfBD1F2Bc3A1F59c494125e07C48F9",
-        usdc_address="0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174",
-        wbtc_address="0x1BFD67037B42Cf73acF2047067bd4F2C47D9BfD6",
-        weth_address="0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619",
-    ),
-    "avalanche:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="0xbabe61887f1de2713c6f97e567623453d3c79f67",
-        fee_receiver_address="0x06534b0BF7Ff378F162d4F348390BDA53b15fA35",
-        usdc_address="0xB97EF9Ef8734C71904D8002F8b6Bc66Dd9c48a6E",
-        wbtc_address="0x50b7545627a5162F82A992c33b87aDc75187B218",
-        weth_address="0x49D5c2BdFfac6CE2BFdB6640F4F80f226bc10bAB",
+        thin_proxy_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
+        fee_receiver_address="0x774D1Dba98cfBD1F2Bc3A1F59c494125e07C48F9",  # burner
     ),
     "gnosis:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="",  # <--- need to deploy sidechain ownership contract  # noqa: E501
-        fee_receiver_address="",  # <--- need to deploy sidechain pool proxy
-        usdc_address="0xDDAfbb505ad214D7b80b1f830fcCc89B60fb7A83",
-        wbtc_address="0x8e5bBbb09Ed1ebdE8674Cda39A0c169401db4252",
-        weth_address="0x6A023CCd1ff6F2045C3309768eAd9E68F978f6e1",
+        thin_proxy_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
+        fee_receiver_address="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
     ),
-    "fantom:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",  # <--- thin proxy  # noqa: E501
-        fee_receiver_address="0x2B039565B2b7a1A9192D4847fbd33B25b836B950",
-        usdc_address="0x04068DA6C83AFCFA0e13ba15A6696662335D5B75",  # <-- multichain usdc  # noqa: E501
-        wbtc_address="0x321162Cd933E2Be498Cd2267a90534A804051b11",  # <-- multichain wbtc  # noqa: E501
-        weth_address="0x74b23882a30290451A17c44f4F05243b6b58C76d",  # <-- multichain weth  # noqa: E501
+    "avax:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
+        fee_receiver_address="0x06534b0BF7Ff378F162d4F348390BDA53b15fA35",  # burner
     ),
-    "celo:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="0x56bc95Ded2BEF162131905dfd600F2b9F1B380a4",  # <-- needs to accept transfer ownership for 0x5277A0226d10392295E8D383E9724D6E416d6e6C  # noqa: E501
-        fee_receiver_address="0x56bc95Ded2BEF162131905dfd600F2b9F1B380a4",  # <-- Thin proxy, needs to be changed!  # noqa: E501
-        usdc_address="0x37f750b7cc259a2f741af45294f6a16572cf5cad",  # <-- wormhole usdc  # noqa: E501
-        wbtc_address="",
-        weth_address="0x66803FB87aBd4aaC3cbB3fAd7C3aa01f6F3FB207",  # <-- wormhole weth  # noqa: E501
+    "ftm:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0xB055EbbAcc8Eefc166c169e9Ce2886D0406aB49b",
+        fee_receiver_address="0x2B039565B2b7a1A9192D4847fbd33B25b836B950",  # burner
     ),
     "kava:mainnet": CurveNetworkSettings(
-        dao_ownership_contract="",
-        fee_receiver_address="",
-        usdc_address="",
-        wbtc_address="",
-        weth_address="",
+        thin_proxy_contract="0x1f0e8445Ebe0D0F60A96A7cd5BB095533cb15B58",
+        fee_receiver_address="0x1f0e8445Ebe0D0F60A96A7cd5BB095533cb15B58",
     ),
-    "moonbeam": CurveNetworkSettings(
-        dao_ownership_contract="",
-        fee_receiver_address="",
-        usdc_address="",
-        wbtc_address="",
-        weth_address="",
+    "celo:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0x56bc95Ded2BEF162131905dfd600F2b9F1B380a4",
+        fee_receiver_address="0x56bc95Ded2BEF162131905dfd600F2b9F1B380a4",
     ),
-    "aurora": CurveNetworkSettings(
-        dao_ownership_contract="",
-        fee_receiver_address="",
-        usdc_address="0xb12bfca5a55806aaf64e99521918a4bf0fc40802",
-        wbtc_address="",
-        weth_address="",
+    "aurora:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
+        fee_receiver_address="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
     ),
-    "mainnet-fork": CurveNetworkSettings(
-        dao_ownership_contract="0x40907540d8a6C65c637785e8f8B742ae6b0b9968",
-        fee_receiver_address="0xeCb456EA5365865EbAb8a2661B0c503410e9B347",
-        usdc_address="0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-        wbtc_address="0x2260fac5e5542a773aa44fbcfedf7c193bc2c599",
-        weth_address="0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
+    "bsc:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="0x98B4029CaBEf7Fd525A36B0BF8555EC1d42ec0B6",
+        fee_receiver_address="0x98B4029CaBEf7Fd525A36B0BF8555EC1d42ec0B6",
+    ),
+    "mantle:mainnet": CurveNetworkSettings(
+        dao_ownership_contract="0xfe87a6CDCa1EEb90987c6A196a1C5F5C76F5F2b0",
+        thin_proxy_contract="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
+        fee_receiver_address="0xf3A431008396df8A8b2DF492C913706BDB0874ef",
+    ),
+    "tron:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="",
+        fee_receiver_address="",
+    ),
+    "zksync:mainnet": CurveNetworkSettings(
+        thin_proxy_contract="",
+        fee_receiver_address="0x4920088D9a5e5De9c098FCA4960d0DA5f4caa4c1",
     ),
 }
-
 
 CURVE_DAO_OWNERSHIP = {
     "agent": "0x40907540d8a6c65c637785e8f8b742ae6b0b9968",
@@ -182,290 +123,32 @@ CURVE_DAO_OWNERSHIP = {
 }
 
 
-def get_tricrypto_usdc_params():
-
-    cg = CoinGeckoAPI()
-    TOKEN_PRICES = []
-    for coin in ["usd-coin", "wrapped-bitcoin", "ethereum"]:
-        TOKEN_PRICES.append(
-            cg.get_price(ids=coin, vs_currencies="usd")[coin]["usd"]
-        )
-    USDC_PRICE = TOKEN_PRICES[0]
-    INITIAL_PRICES = [int(p / USDC_PRICE) * 10**18 for p in TOKEN_PRICES[1:]]
-    MA_TIME_SECONDS = 600  # seconds
-    return {
-        "name": "TricryptoUSDC",
-        "symbol": "crvUSDCWBTCWETH",
-        "coins": [],
-        "implementation_index": 0,
-        "A": 1707629,
-        "gamma": 11809167828997,
-        "mid_fee": 3000000,
-        "out_fee": 30000000,
-        "allowed_extra_profit": 2000000000000,
-        "fee_gamma": 500000000000000,
-        "adjustment_step": 490000000000000,
-        "ma_time": int(MA_TIME_SECONDS / math.log(2)),
-        "initial_prices": INITIAL_PRICES,
-    }
-
-
-def test_deployment(pool, coins, fee_receiver, account):
-
-    PARAMS = get_tricrypto_usdc_params()
-
-    logger.info(
-        "------------ TEST IF CONTRACT WORKS AS INTENDED IN PROD ----------------"  # noqa: E501
-    )
-
-    for coin in coins:
-
-        coin_contract = Contract(coin)
-        bal = coin_contract.balanceOf(account)
-        assert bal > 0, "Not enough coins!"
-
-        if coin_contract.allowance(account, pool) > bal:
-            continue
-
-        coin_name = coin_contract.name()
-        logger.info(f"Approve pool to spend deployer's {coin_name}:")
-
-        coin_contract.approve(
-            pool, 2**256 - 1, sender=account, **_get_tx_params()
-        )
-
-    logger.info("------------------------------ Add liquidity")
-
-    logger.info("Deposit WETH with other tokens:")
-    tokens_to_add = get_deposit_amounts(
-        DOLLAR_VALUE_OF_TOKENS_TO_DEPOSIT, PARAMS["initial_prices"], coins
-    )
-
-    logger.info(f"Add {tokens_to_add} tokens to deployed pool: ")
-
-    tx = pool.add_liquidity(
-        tokens_to_add,
-        0,
-        False,
-        sender=account,
-        **_get_tx_params(),
-    )
-    d_tokens = tx.return_value
-    assert pool.balanceOf(account) == pool.totalSupply() == d_tokens
-    logger.info(f"Received {d_tokens} number of LP Tokens.")
-
-    logger.info("Deposit ETH with other tokens:")
-    tx = pool.add_liquidity(
-        tokens_to_add,
-        0,
-        True,
-        sender=account,
-        value=tokens_to_add[2],
-        **_get_tx_params(),
-    )
-    d_tokens = tx.return_value
-    assert d_tokens > 0
-    logger.info(f"Received {d_tokens} number of LP Tokens.")
-
-    logger.info("------------------------------ Exchange")
-
-    amt_usdc_in = 10 * 10 ** project.ERC20Mock.at(coins[0]).decimals()
-    logger.info(f"Test exchange_underlying of {amt_usdc_in} USDC -> ETH:")
-    tx = pool.exchange_underlying(
-        0,
-        2,
-        amt_usdc_in,
-        0,
-        sender=account,
-        **_get_tx_params(),
-    )
-    dy_eth = tx.events.filter(pool.TokenExchange)[
-        0
-    ].tokens_bought  # return_value is broken in ape somehow
-    assert dy_eth > 0
-    logger.info(f"Received {dy_eth} ETH")
-
-    logger.info(f"Test exchange_underlying of {dy_eth} ETH -> USDC:")
-    tx = pool.exchange_underlying(
-        2,
-        0,
-        dy_eth,
-        0,
-        sender=account,
-        value=dy_eth,
-        **_get_tx_params(),
-    )
-    dy_usdc = tx.events.filter(pool.TokenExchange)[0].tokens_bought
-    assert dy_usdc > 0
-    logger.info(f"Received {dy_usdc} USDC")
-
-    logger.info(f"Test exchange of {dy_usdc} USDC -> WBTC:")
-    tx = pool.exchange(
-        0,
-        1,
-        dy_usdc * 2,
-        0,
-        sender=account,
-        **_get_tx_params(),
-    )
-    dy_wbtc = tx.events.filter(pool.TokenExchange)[0].tokens_bought
-    assert dy_wbtc > 0
-    logger.info(f"Received {dy_wbtc} WBTC")
-
-    logger.info("------------------------------ Remove Liquidity in one coin")
-
-    bal = pool.balanceOf(account)
-    amt_to_remove = int(bal / 4)
-    logger.info(f"Remove {amt_to_remove} liquidity in native token (ETH):")
-    tx = pool.remove_liquidity_one_coin(
-        amt_to_remove,
-        2,
-        0,
-        True,
-        sender=account,
-        **_get_tx_params(),
-    )
-    dy_eth = tx.events.filter(pool.RemoveLiquidityOne)[0].coin_amount
-    assert dy_eth > 0
-    logger.info(f"Removed {dy_eth} of ETH.")
-
-    for coin_id, coin in enumerate(coins):
-
-        bal = pool.balanceOf(account)
-        coin_contract = project.ERC20Mock.at(coin)
-        coin_name = coin_contract.name()
-        coin_balance = coin_contract.balanceOf(account)
-
-        logger.info(f"Remove {int(bal/4)} liquidity in {coin_name}:")
-        tx = pool.remove_liquidity_one_coin(
-            int(bal / 4),
-            coin_id,
-            0,
-            False,
-            sender=account,
-            **_get_tx_params(),
-        )  # noqa: E501
-
-        dy_coin = tx.events.filter(pool.RemoveLiquidityOne)[0].coin_amount
-        assert dy_coin > 0
-        assert coin_contract.balanceOf(account) == coin_balance + dy_coin
-        logger.info(f"Removed {dy_coin} of {coin_name}.")
-
-    logger.info(
-        "------------------------------ Remove liquidity proportionally"
-    )
-
-    bal = pool.balanceOf(account)
-    logger.info(
-        f"Remove {int(bal/4)} amount of liquidity proportionally (with native ETH):"  # noqa: E501
-    )
-    tx = pool.remove_liquidity(
-        int(bal / 4),
-        [0, 0, 0],
-        True,
-        sender=account,
-        **_get_tx_params(),
-    )
-    dy_tokens = tx.events.filter(pool.RemoveLiquidity)[0].token_amounts
-    for tkn_amt in dy_tokens:
-        assert tkn_amt > 0
-
-    logger.info(f"Removed {dy_tokens} of liquidity.")
-
-    logger.info(
-        f"Remove {int(bal/4)} amount of liquidity proportionally (with native ETH):"  # noqa: E501
-    )
-    tx = pool.remove_liquidity(
-        int(bal / 4),
-        [0, 0, 0],
-        False,
-        sender=account,
-        **_get_tx_params(),
-    )
-    dy_tokens = tx.return_value
-
-    for tkn_amt in dy_tokens:
-        assert tkn_amt > 0
-
-    logger.info("Successfully tested deployment!")
-
-
-def deploy_amm_factory(
-    account, fee_receiver, weth, network, deployed_contracts={}
+def get_create2_deployment_address(
+    compiled_bytecode,
+    abi_encoded_ctor,
+    salt,
+    create2deployer,
+    blueprint=False,
+    blueprint_preamble=b"\xFE\x71\x00",
 ):
-
-    if "math" not in deployed_contracts:
-        logger.info("Deploying math contract:")
-        math_contract = account.deploy(
-            project.CurveCryptoMathOptimized3, **_get_tx_params()
-        )
-    else:
-        math_contract = project.CurveCryptoMathOptimized3.at(
-            deployed_contracts["math"]
-        )
-
-    if "views" not in deployed_contracts:
-        logger.info("Deploying views contract:")
-        views_contract = account.deploy(
-            project.CurveCryptoViews3Optimized, **_get_tx_params()
-        )
-    else:
-        views_contract = project.CurveCryptoViews3Optimized.at(
-            deployed_contracts["views"]
+    deployment_bytecode = compiled_bytecode + abi_encoded_ctor
+    if blueprint:
+        # Add blueprint preamble to disable calling the contract:
+        blueprint_bytecode = blueprint_preamble + deployment_bytecode
+        # Add code for blueprint deployment:
+        len_blueprint_bytecode = len(blueprint_bytecode).to_bytes(2, "big")
+        deployment_bytecode = (
+            b"\x61"
+            + len_blueprint_bytecode
+            + b"\x3d\x81\x60\x0a\x3d\x39\xf3"
+            + blueprint_bytecode
         )
 
-    if "amm_impl" not in deployed_contracts:
-        logger.info("Deploying AMM blueprint contract:")
-        amm_impl = deploy_blueprint(
-            project.CurveTricryptoOptimizedWETH, account
-        )
-    else:
-        amm_impl = project.CurveTricryptoOptimizedWETH.at(
-            deployed_contracts["amm_impl"]
-        )
-
-    if "factory" not in deployed_contracts:
-        logger.info("Deploy factory:")
-
-        if any([_n in network for _n in ["arbitrum", "optimism"]]):
-            factory = project.CurveL2TricryptoFactory.deploy(
-                fee_receiver,
-                account.address,
-                sender=account,
-                **_get_tx_params(),
-            )
-        elif "ethereum" in network:
-            factory = project.CurveTricryptoFactory.deploy(
-                fee_receiver,
-                account.address,
-                sender=account,
-                **_get_tx_params(),
-            )
-        else:
-            raise NotImplementedError
-
-        constructor_args = [fee_receiver, account.address, weth]
-        logger.info(
-            f"Constructor args: {encode(['address', 'address', 'address'], constructor_args).hex()}\n"  # noqa: E501
-        )
-    else:
-        factory = project.CurveTricryptoFactory.at(
-            deployed_contracts["factory"]
-        )
-
-    logger.info("Set Pool Implementation:")
-    factory.set_pool_implementation(
-        amm_impl, 0, sender=account, **_get_tx_params()
+    return (
+        create2deployer.computeAddress(salt, keccak(deployment_bytecode)),
+        deployment_bytecode,
     )
 
-    logger.info("Set Views implementation:")
-    factory.set_views_implementation(
-        views_contract, sender=account, **_get_tx_params()
-    )
 
-    logger.info("Set Math implementation:")
-    factory.set_math_implementation(
-        math_contract, sender=account, **_get_tx_params()
-    )
-
-    return factory
+def deploy_via_create2_factory(deployment_bytecode, salt, create2deployer):
+    create2deployer.deploy(0, salt, deployment_bytecode)
